Fix forEachLayerAtPixel and improve class hierarchy
